requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.3
PyMuPDF==1.24.10
